requirements.txt
python-dotenv~=1.0.1
PyYAML~=6.0.1
flake8~=7.0.0
pymongo~=4.6.2
fastapi~=0.110.0
uvicorn~=0.28.0
openai~=1.13.3
transformers~=4.38.2
torch~=2.2.1
fastapi~=0.110.0
pydantic~=2.6.3
pillow~=10.2.0
names~=0.3.0
stability_sdk~=1.0.0